Adjusted Operation
Okay I flushed it out a tad bit more. Before, the code would just respit out the buffer, so 26 * would display 2626. Now 26* will display 0. However, once an operation is selected, the next doIt (the button pressed after the operation) will call on NextOperation's DoIt. 
A possible solution is to change the next state from doing an operation to the operation, adjust the needed variables, then once all that is finished, have the next state go back to Value1, which is what handles most inputs. With this said, value2 may not even be used.